+++ Frends.SFTP.DownloadFiles/Frends.SFTP.DownloadFiles/Definitions/FileTransporter.cs
@@ -74,7 +74,12 @@ internal class FileTransporter
                 client.ConnectionInfo.KeyExchangeAlgorithms.Remove("curve25519-sha256");
                 client.ConnectionInfo.KeyExchangeAlgorithms.Remove("curve25519-sha256@libssh.org");
 
+                // Force task to use ssh-rsa algotrithm because otherwise ED25519 is used.
+                client.ConnectionInfo.HostKeyAlgorithms.Clear();
+                client.ConnectionInfo.HostKeyAlgorithms.Add("ssh-rsa", (data) => { return new KeyHostAlgorithm("ssh-rsa", new RsaKey(), data); });
+
                 var expectedServerFingerprint = _batchContext.Connection.ServerFingerPrint;
+
                 // Check the fingerprint of the server if given.
                 if (!string.IsNullOrEmpty(expectedServerFingerprint))
                 {
@@ -243,7 +248,7 @@ internal class FileTransporter
     private void CheckServerFingerprint(SftpClient client, string expectedServerFingerprint)
     {
         var userResultMessage = "";
-        
+
         client.HostKeyReceived += delegate (object sender, HostKeyEventArgs e)
         {
             if (Util.IsMD5(expectedServerFingerprint.Replace(":", "").Replace("-", "")))
@@ -252,7 +257,7 @@ internal class FileTransporter
                 {
                     var serverFingerprint = BitConverter.ToString(e.FingerPrint).Replace("-", "").Replace(":", "");
                     
-                    e.CanTrust = expectedServerFingerprint == serverFingerprint;
+                    e.CanTrust = expectedServerFingerprint.ToLower() == serverFingerprint.ToLower();
                     if (!e.CanTrust)
                         userResultMessage = $"Can't trust SFTP server. The server fingerprint does not match. " +
                                 $"Expected fingerprint: '{expectedServerFingerprint}', but was: '{serverFingerprint}'.";
@@ -268,33 +273,15 @@ internal class FileTransporter
                 }
 
             }
-            else if (Util.IsSha1(expectedServerFingerprint))
-            {
-                using (var sha1 = SHA1.Create())
-                {
-                    var hash = sha1.ComputeHash(e.HostKey);
-                    var sha1Fingerprint = string.Concat(hash.Select(b => b.ToString("x2")));
-                    e.CanTrust = (sha1Fingerprint == expectedServerFingerprint);
-                    if (!e.CanTrust)
-                        userResultMessage = $"Can't trust SFTP server. The server fingerprint does not match. " +
-                                            $"Expected fingerprint: '{expectedServerFingerprint}', but was: '{sha1Fingerprint}'.";
-                }
-            }
             else if (Util.IsSha256(expectedServerFingerprint))
             {
-                // Try with Sha256 algorithm
-                client.ConnectionInfo.HostKeyAlgorithms.Clear();
-                client.ConnectionInfo.HostKeyAlgorithms.Add("ssh-rsa", (data) => { return new KeyHostAlgorithm("ssh-rsa", new RsaKey(), data); });
-                
                 if (Util.TryConvertHexStringToHex(expectedServerFingerprint))
                 {
                     using (SHA256 mySHA256 = SHA256.Create())
                     {
                         var sha256Fingerprint = Util.ToHex(mySHA256.ComputeHash(e.HostKey));
-                        // Set the userResultMessage in case checking server fingerprint failes.
 
                         e.CanTrust = (sha256Fingerprint == expectedServerFingerprint);
-                        var test = e.CanTrust;
                         if (!e.CanTrust)
                             userResultMessage = $"Can't trust SFTP server. The server fingerprint does not match. " +
                                                 $"Expected fingerprint: '{expectedServerFingerprint}', but was: '{sha256Fingerprint}'.";
@@ -305,18 +292,21 @@ internal class FileTransporter
                     using (SHA256 mySHA256 = SHA256.Create())
                     {
                         var sha256Fingerprint = Convert.ToBase64String(mySHA256.ComputeHash(e.HostKey));
-                        e.CanTrust = (sha256Fingerprint == expectedServerFingerprint);
+                        e.CanTrust = (sha256Fingerprint == expectedServerFingerprint || sha256Fingerprint.Replace("=", "") == expectedServerFingerprint);
                         if (!e.CanTrust)
                             userResultMessage = $"Can't trust SFTP server. The server fingerprint does not match. " +
                                                 $"Expected fingerprint: '{expectedServerFingerprint}', but was: '{sha256Fingerprint}'.";
                     }
                 }
             }
-        else { e.CanTrust = false; }
+            else 
+            {
+                userResultMessage = "Expected server fingerprint was given in unsupported format.";
+                e.CanTrust = false; 
+            }
 
-        if (!e.CanTrust)
-            _logger.NotifyError(_batchContext, userResultMessage, new SshConnectionException());
-
+            if (!e.CanTrust)
+                _logger.NotifyError(_batchContext, userResultMessage, new SshConnectionException());
         };
     }